Change visitId to Long

+++ src/main/java/org/lamisplus/modules/pharmacy/domain/dto/DrugOrderDTO.java
@@ -57,6 +57,5 @@ public class DrugOrderDTO {
 
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd@HH:mm:ss")
     private LocalDateTime dateTimeDispensed;
-
-    private Integer visitId;
+    private Long visitId;
 }
